+++ CHANGES.rst
@@ -30,6 +30,7 @@ Improvements
 - Allow changing review tags in the editing timeline (:issue:`7133`, :pr:`7134`)
 - Add an option to request changes in bulk on the editable list (:issue:`7062`,
   :pr:`7100`)
+- Clone persons settings when cloning an event (:pr:`7158`)
 
 Bugfixes
 ^^^^^^^^